renamed eventPage to utils
