+++ client/background/utils.js
